refactor(agent): isolate pi-specific chat helpers
